lottery/requirements.txt
torch==2.5.1
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
pyyaml>=6.0.1
scikit-learn>=1.3.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
gymnasium>=0.28.1
lightgbm>=3.3.5
xgboost>=1.7.4
psutil>=5.9.0
optuna>=3.1.0
pymc>=5.0.0
arviz>=0.11.0
bayesian-optimization>=1.2.0 